Refactor(frontend): Redux re-organization
Reorganized Redux in order to expose all types and such through the
standard pattern of importing the directory and exporting all
public/library facing features from index.ts because it's a library.
