+++ frontend/redux/ThemeAction.ts
@@ -1,4 +1,4 @@
-import { Store } from "./UserReducer";
+import { Store } from ".";
 
 export const updateTheme = (theme: Store["theme"]) => ({
   type: "UPDATE_THEME",
