+++ frontend/redux/UserAction.ts
@@ -1,4 +1,4 @@
-import { Store } from "./UserReducer";
+import { Store } from ".";
 
 export const updateUser = (user: Store["user"]) => ({
   type: "UPDATE_USER",
